+++ cnd/src/config/mod.rs
@@ -5,10 +5,17 @@ pub mod settings;
 use crate::swap_protocols::ledger::ethereum;
 use libp2p::Multiaddr;
 use serde::{Deserialize, Serialize};
-use std::{net::IpAddr, path::PathBuf};
+use std::{
+    net::{IpAddr, Ipv4Addr, SocketAddr},
+    path::PathBuf,
+};
 
 pub use self::{file::File, settings::Settings};
 use reqwest::Url;
+
+lazy_static::lazy_static! {
+    pub static ref LND_SOCKET: SocketAddr = SocketAddr::new(IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)), 8080);
+}
 
 #[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
 pub struct Data {
@@ -18,12 +25,6 @@ pub struct Data {
 #[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
 pub struct Network {
     pub listen: Vec<Multiaddr>,
-}
-
-#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
-pub struct Socket {
-    pub address: IpAddr,
-    pub port: u16,
 }
 
 #[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
@@ -91,6 +92,40 @@ pub struct Parity {
     pub node_url: reqwest::Url,
 }
 
+#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
+pub struct Lightning {
+    pub network: bitcoin::Network,
+    pub lnd: Option<Lnd>,
+}
+
+impl Default for Lightning {
+    fn default() -> Self {
+        Self {
+            network: bitcoin::Network::Regtest,
+            lnd: Some(Lnd::default()),
+        }
+    }
+}
+
+#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
+pub struct Lnd {
+    pub rest_api_socket: Option<SocketAddr>,
+    pub dir: Option<PathBuf>,
+}
+
+impl Default for Lnd {
+    fn default() -> Self {
+        Self {
+            rest_api_socket: Some(*LND_SOCKET),
+            dir: Some(default_lnd_dir()),
+        }
+    }
+}
+
+fn default_lnd_dir() -> PathBuf {
+    crate::lnd_dir().expect("no home directory")
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -122,6 +157,82 @@ mod tests {
             .into_iter()
             .map(toml::from_str)
             .collect::<Result<Vec<Network>, toml::de::Error>>()
+            .unwrap();
+
+        assert_eq!(actual, expected);
+    }
+
+    #[test]
+    fn lnd_deserializes_correctly() {
+        let file_contents = vec![
+            r#"
+            rest_api_socket = "127.0.0.1:8080"
+            dir = "~/.local/share/comit/lnd"
+            "#,
+            r#"
+            rest_api_socket = "127.0.0.1:8080"
+            "#,
+            r#"
+            dir = "~/.local/share/comit/lnd"
+            "#,
+        ];
+
+        let expected = vec![
+            Lnd {
+                rest_api_socket: Some(*LND_SOCKET),
+                dir: Some(PathBuf::from("~/.local/share/comit/lnd")),
+            },
+            Lnd {
+                rest_api_socket: Some(*LND_SOCKET),
+                dir: None,
+            },
+            Lnd {
+                rest_api_socket: None,
+                dir: Some(PathBuf::from("~/.local/share/comit/lnd")),
+            },
+        ];
+
+        let actual = file_contents
+            .into_iter()
+            .map(toml::from_str)
+            .collect::<Result<Vec<Lnd>, toml::de::Error>>()
+            .unwrap();
+
+        assert_eq!(actual, expected);
+    }
+
+    #[test]
+    fn lightning_deserializes_correctly() {
+        let file_contents = vec![
+            r#"
+            network = "regtest"
+            "#,
+            r#"
+            network = "regtest"
+            [lnd]
+            rest_api_socket = "127.0.0.1:8080"
+            dir = "/path/to/lnd"
+            "#,
+        ];
+
+        let expected = vec![
+            Lightning {
+                network: bitcoin::Network::Regtest,
+                lnd: None,
+            },
+            Lightning {
+                network: bitcoin::Network::Regtest,
+                lnd: Some(Lnd {
+                    rest_api_socket: Some(*LND_SOCKET),
+                    dir: Some(PathBuf::from("/path/to/lnd")),
+                }),
+            },
+        ];
+
+        let actual = file_contents
+            .into_iter()
+            .map(toml::from_str)
+            .collect::<Result<Vec<Lightning>, toml::de::Error>>()
             .unwrap();
 
         assert_eq!(actual, expected);