+++ cnd/src/lib.rs
@@ -50,8 +50,11 @@ pub mod swap_protocols;
 pub mod timestamp;
 
 use anyhow::Context;
-use directories::ProjectDirs;
-use std::path::{Path, PathBuf};
+use directories::{ProjectDirs, UserDirs};
+use std::{
+    env,
+    path::{Path, PathBuf},
+};
 
 /// Define domain specific terms using identity module so that we can refer to
 /// things in an ergonomic fashion e.g., `identity::Bitcoin`.
@@ -89,6 +92,27 @@ pub fn default_config_path() -> anyhow::Result<PathBuf> {
 // OSX: /Users/<user>/Library/Application Support/comit/
 pub fn data_dir() -> Option<PathBuf> {
     ProjectDirs::from("", "", "comit").map(|proj_dirs| proj_dirs.data_dir().to_path_buf())
+}
+
+/// Returns `/Users/[username]/Library/Application Support/Lnd/`.
+/// exists.
+#[cfg(target_os = "macos")]
+pub fn lnd_default_dir() -> Option<PathBuf> {
+    ProjectDirs::from("", "", "Lnd").map(|proj_dirs| proj_dirs.data_dir().to_path_buf())
+}
+
+/// Returns `~/.lnd` if $HOME exists.
+#[cfg(target_os = "linux")]
+pub fn lnd_default_dir() -> Option<PathBuf> {
+    UserDirs::new().map(|d| d.home_dir().to_path_buf().join(".lnd"))
+}
+
+/// Returns the directory used by lnd.
+pub fn lnd_dir() -> Option<PathBuf> {
+    if let Ok(dir) = env::var("LND_DIR") {
+        return Some(PathBuf::from(dir));
+    }
+    lnd_default_dir()
 }
 
 pub type Never = std::convert::Infallible;